PostgreSQL 18 no longer supports UNLOGGED partitioned tables.
This commit adds a warning when an UNLOGGED partitioned table is created,
and skips the UNLOGGED modifier to ensure that the table is created as a
regular (LOGGED) partitioned table.

It also silences this warning in existing tests where UNLOGGED + PARTITION
is not the focus of the test, and adds a new test to explicitly assert that
the warning is emitted in PostgreSQL 18 and later.

One of the errors addressed by this commit:

  $ ARCONN=postgresql bin/test test/cases/adapters/postgresql/partitions_test.rb -n test_partitions_table_exists
  ...
  ActiveRecord::StatementInvalid: PG::FeatureNotSupported: ERROR:
  partitioned tables cannot be unlogged
  ...

Without this commit, the generated SQL was:

  CREATE UNLOGGED TABLE "partitioned_events" ("issued_at" timestamp) PARTITION BY RANGE (issued_at)

With this commit, it becomes:

  CREATE TABLE "partitioned_events" ("issued_at" timestamp) PARTITION BY RANGE (issued_at)

Reference: PostgreSQL 18 release notes
https://www.postgresql.org/docs/18/release-18.html

> Disallow unlogged partitioned tables (Michael Paquier)
>
> Previously ALTER TABLE SET [UN]LOGGED did nothing,
> and the creation of an unlogged partitioned table did not cause its children to be unlogged.

Source: https://git.postgresql.org/gitweb/?p=postgresql.git;a=commitdiff;h=e2bab2d79